+++ src/midi/index.ts
@@ -4,3 +4,4 @@ export * from "./errors";
 export * from "./ports";
 export * from "./reassembly";
 export * from "./request-response";
+export * from "./wire-log";